move validation of a ray desc to be shared
